- Renamed data to store - Added Auto check for Js files - Added new command: ```lint:watch```
