+++ components/SignInForm/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import AuthFields from '../AuthFields';
 import validate from '../AuthFields/validation';
-import connect from './data';
+import connect from './store';
 
 class SignInForm extends React.Component {
   static propTypes = {
