+++ next.config.js
@@ -8,6 +8,15 @@ module.exports = {
 
     config.plugins.push(new IgnorePlugin(/^\.\/locale$/, /moment$/));
 
+    if (dev) {
+      config.module.rules.push({
+        test: /\.(jsx?|gql|graphql)$/,
+        loader: 'eslint-loader',
+        exclude: ['/node_modules/', '/.next/', '/helper_scripts/'],
+        enforce: 'pre'
+      });
+    }
+
     if (process.env.ANALYZE_BUILD) {
       config.plugins.push(
         new BundleAnalyzerPlugin({
