+++ package.json
@@ -40,7 +40,7 @@
     "start": "NODE_ENV=production node server.js",
     "start:next": "next start",
     "lint": "eslint --fix --ext .gql --ext .graphql --ext .js pages components containers libraries server helper_scripts",
-    "lint:watch": "esw -w --fix --ext .gql --ext .graphql --ext .js pages components containers libraries server helper_scripts",
+    "lint:watch": "esw -w --fix --ext .gql --ext .graphql components libraries server",
     "analyze": "ANALYZE_BUILD=true npm run build",
     "setup": "node ./helper_scripts/CL_commands/setup.js && npm run setup:after",
     "setup:after": "npm run build",
@@ -89,12 +89,13 @@
     "eslint": "3.19.0",
     "eslint-config-airbnb": "15.1.0",
     "eslint-config-prettier": "2.3.0",
+    "eslint-loader": "1.9.0",
     "eslint-plugin-graphql": "1.3.0",
     "eslint-plugin-import": "2.7.0",
     "eslint-plugin-jsx-a11y": "5.1.1",
     "eslint-plugin-prettier": "2.1.2",
-    "eslint-watch": "3.1.2",
     "eslint-plugin-react": "7.2.0",
+    "eslint-watch": "3.1.2",
     "figlet": "1.2.0",
     "graphql-cli": "1.0.0-beta.4",
     "graphql-cli-voyager": "0.1.2",
